+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Internal refactoring to reduce required memory usage
+* New tag filter: `-f ~k` to filter on keys which match a regex
 
 ## v0.4.0 (2023-08-17)